create component for app bar - Tucker
