+++ src/App.js
@@ -21,44 +21,12 @@ import BedIcon from '@mui/icons-material/Bed';
 import VisibilityIcon from '@mui/icons-material/Visibility';
 import DateRangeIcon from '@mui/icons-material/DateRange';
 import AddchartIcon from '@mui/icons-material/Addchart';
+import AppBarComponent from './Components/AppBarComponent';
 
 const config = require('./config');
 
-const links = [
-  {
-    name: "Home",
-    location: "/",
-  },
-  {
-    name: "Import Room",
-    location: "/importRoomType"
-  },
-  {
-    name: "Import View",
-    location: "/importViewType"
-  },
-  {
-    name: "Import Points",
-    location: "/importPointsTable"
-  },
-  {
-    name: "Import Point Block",
-    location: "/importPointBlock"
-  },
-]
-
 function App() {
   const [resorts, setResorts] = useState([]);
-  const [anchorEl, setAnchorEl] = useState(null);
-  const open = Boolean(anchorEl);
-
-  const handleMenuClick = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
 
   useEffect(() => {
     axios.get(`${config.api.protocol}://${config.api.host}/dvc-calc-api/resorts`).then(resp => {
@@ -69,76 +37,7 @@ function App() {
   return (
     <Router>
       <div className="App">
-        <AppBar position='static'>
-
-
-          <Toolbar>
-            <IconButton
-              size="large"
-              edge="start"
-              color="inherit"
-              aria-label="menu"
-              sx={{ mr: 2 }}
-              onClick={handleMenuClick}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Drawer
-              id="sideMenu"
-              anchorEl={anchorEl}
-              open={open}
-              onClose={handleClose}
-            >
-              <List>
-                  <ListItem component={Link} href="/">
-                    <ListItemButton>
-                      <ListItemIcon>
-                        <CalculateIcon></CalculateIcon>
-                      </ListItemIcon>
-                      <ListItemText primary="DVC Calculator" />
-                    </ListItemButton>
-                  </ListItem>
-                  <ListItem component={Link} href="/importRoomType">
-                    <ListItemButton>
-                      <ListItemIcon>
-                        <BedIcon></BedIcon>
-                      </ListItemIcon>
-                      <ListItemText primary="Import Room Type" />
-                    </ListItemButton>
-                  </ListItem>
-                  <ListItem component={Link} href="/importViewType">
-                    <ListItemButton>
-                      <ListItemIcon>
-                        <VisibilityIcon></VisibilityIcon>
-                      </ListItemIcon>
-                      <ListItemText primary="Import View Type" />
-                    </ListItemButton>
-                  </ListItem>
-                  <ListItem component={Link} href="/importPointBlock">
-                    <ListItemButton>
-                      <ListItemIcon>
-                        <DateRangeIcon></DateRangeIcon>
-                      </ListItemIcon>
-                      <ListItemText primary="Import Point Block" />
-                    </ListItemButton>
-                  </ListItem>
-                  <ListItem component={Link} href="/importPointsTable">
-                    <ListItemButton>
-                      <ListItemIcon>
-                        <AddchartIcon></AddchartIcon>
-                      </ListItemIcon>
-                      <ListItemText primary="Import Point Table" />
-                    </ListItemButton>
-                  </ListItem>
-              </List>
-            </Drawer>
-            <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-              News
-            </Typography>
-            <Button color="inherit">Login</Button>
-          </Toolbar>
-
-        </AppBar>
+        <AppBarComponent />
         <Routes>
           <Route path="/" element={<DVCCalculatorComponent resorts={resorts} />} />
           <Route path="/importPoints" element={<ImportPointsComponent resorts={resorts} />} />
